Update LICENSE so GitHub recognizes it?
For some reason, the MIT LICENSE of the repo is not spotted by GitHub, and this cause the package to appear unlicensed in SwiftPackageIndex.

I'll try to update slightly the file to see if GitHub finally matches it up. Unfortunately, I need to do this on `main`.

I won't change the license's conditions, and the package will stay under the MIT License.