Allow the dependency chain to be updated at any time.
This commit removes the restriction which prohibited dependencies values to be modified once the chain of environment was accessed for the first time.

This safeguard was installed because environments did retrieve dependencies from their parent only once, on first access (for performance reasons), without any updating mechanism.

This restriction is now lifted and environments are now spotting when their parent's dependencies have potentially changed, and update accordingly.